Add hyperv module to hypervisor crate
This patch introduces the hyperv module to the hypervisor
crate. Along with KVM, the plan is to add another hypervisor support.
We will have similar implementation of all the traits defined in the
hypervisor crate. This patch adds the skeleton of the module that
deals with the support of Microsoft hyperv hypervisor.

Signed-off-by: Muminul Islam <muislam@microsoft.com>