Hack: revive VcpuRun trait
For VmExit::IoIn/IoOut/MmioRead/MmioWrite, it will need to take a
reference to a slice named "data". On KVM the slice is derived from
KVM's shared mapping between the VMM and the kernel. That has some
advantages:
  1. The lifetime constraint on the reference is easy to sastify
  2. There is no extra step needed to copy to / from a buffer to be
     consume for the kernel

On Hyper-V, we don't have that luxury. We will need to:
  1. Maintain a per-vcpu buffer that can be used to pass data between
     our code and other code
  2. Do more complex state tracking to pass data from / to the
     hypervisor using hypercall or whatever

Seems that the fastest way to get things working for now is to revive
VcpuRun trait.

Signed-off-by: Wei Liu <liuwe@microsoft.com>