hypervisor: mshv: Implement a software based TLB
A very naive implementation, that basically maps gva to gpa.
This is needed by the instruction emulator, as it only passes linear
addresses down to the platform emulator. Those addresses need to be
translated to physical ones, and this is what this software TLB does.

The TLB is populated at EPT violation VM-Exit, when the hypervisor
forwards the VMCS information that contains the guest gva to gpa
translation. Our software TLB basically caches that piece of information
and allows the emulator to transparently retrieve it.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>
Signed-off-by: Wei Liu <liuwe@microsoft.com>