mshv: Pin mshv to tag: v0.1.1
Instead of using main branch, use a tag for
mshv-{ioctls,bindings} crate. This makes upsteam changes easier to make.
The tag can be updated when a breaking change occurs.

Patch vfio crate to point to a forked repo on github. This is
a temporary measure, to get around the dependency on upstream vfio,
which doesn't yet point to the correct mshv tag.

Signed-off-by: Nuno Das Neves <nudasnev@microsoft.com>