performance-metrics: increase setup time
The existing Setup/cleanup time is not sufficient for
virtio_net_latency_us. It times out on bare metal and L1VH scenarios.
Even in the nested scenario, the test passes by a very narrow margin
and can fail sometimes.

Most of this time is taken by guest boot. The goal of
virtio_net_latency_us is not to measure guest boot. So, it is okay to
increase the timeout here and unblock the test case so that we are
able to obtain network latency numbers.

Increase the timeout to 30.

Signed-off-by: Anirudh Rayabharam <anrayabh@microsoft.com>